Fixed an incorrect return type in Cell class, added parseCell(..) method to XMLLoadParser

+++ GroupProject/src/gameLogic/Cell.java
@@ -41,7 +41,7 @@ public class Cell implements Serializable {
 
 	}
 
-	public double getBatteryLife(){
+	public int getBatteryLife(){	
 		return batteryLife;
 	}
